docs: Add user and developer documentation about Chamelium support
This introduces plain-text documentation about the Chamelium aimed at
users who wish to deploy the platform, as well as developers who wish
to work on improving IGT support for it.

Given the contents of this documentation, it felt more relevant to make
it part of the tree instead of the API reference.

Signed-off-by: Paul Kocialkowski <paul.kocialkowski@linux.intel.com>
(with one small typo fixed)
Reviewed-by: Lyude Paul <lyude@redhat.com>