tools_test: Clean up and fix sysfs_l3_parity
Multiple misunderstandings of the expectations of the test and some
missed parts of the shell-to-c conversion caused a couple of issues to
remain.

First, we need to actually disable a subbank before we check that a
subbank is disabled (invoke the tool with -d).

Second, the original pipe to wc -l was to check that the tool prints
only one line, not that it prints "at least" a line. Modify the last
check to verify that an "is disabled" text is _not_ printed.

v2: Add a TODO comment

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=101650
Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Reviewed-by: Abdiel Janulgue <abdiel.janulgue@linux.intel.com>