intel_aubdump: Support I915_EXEC_BATCH_FIRST.
The batch is now on the other end of the list.

Fixes issues where modern Mesa and modern kernels together start putting
the batch at the front of the list, and intel_aubdump looks for it at
the end of the list, causing it to interpret some other buffer as the
batch.  Then AubLoad or aubinator see bogus data like 1.0 float as a
GPU command and get grumpy.