igt/gem_exec_schedule: Fix up too deep reorder-wide()
Like wide(), reorder-wide() didn't check the ring size before flooding
and so would exhaust its available space and block (causing a GPU hang
to recover), thus failing the test. Also, since we use a new context for
each iteration of the test, the available ring space is reduced (due to
the overhead in setting up that context via the ring), so make sure that
when we measure the available ring space we also use a new context.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>