tests/kms_atomic_transition: Do not update unbound planes.
kms_atomic_transition was updating already disabled planes and committing
them nonblockingly. This results in sporadic -EBUSY failures because
planes that are unbound have their own timeline.

The solution is not unbinding already unbound planes, making the test
pass. There was also a related kernel bug causing failures in the same
way, but that is now fixed.

While at it, only check for fd completion in the nonblocking case,
in the blocking case it's theoretically possible the commit completes
before we check the that the commit didn't complete, most probably when
the test is run in a debugger.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=102671
Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>