tests/kms_concurrent: Fix testcase fails.
igt_pipe_refresh was added for the forked helper, we could just call it
from the fork to prevent races against MODE_ID.

The static 1024x768 mode is allocated on the stack and returned, which will
result in a garbage mode. Make the fallback mode global to prevent this.

igt_output_set_pipe is unneeded, override_mode is enough to force the
right mode.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=103342
Reviewed-by: Mika Kahola <mika.kahola@intel.com>