tests/gem_eio: Nerf in-flight-suspend
Use TEST_NONE instead of TEST_DEVICES to prevent a machine death that
happens on a particular model of SNB (2600 is affected, 2520m is
not). Reset is unreliable, but the exact setup to trigger the death
and how to work around it are not found at this time. There is some
kind of a race lurking, and this commit is a workaround that avoids
it, leaving the test still exercising some of the codepaths.

References: https://intel-gfx-ci.01.org/tree/drm-tip/igt@gem_eio@in-flight-suspend.html
References: https://bugs.freedesktop.org/show_bug.cgi?id=103289
Suggested-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Petri Latvala <petri.latvala@intel.com>
CC: Daniel Vetter <daniel.vetter@ffwll.ch>
CC: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
CC: "Lofstedt, Marta" <marta.lofstedt@intel.com>
CC: Martin Peres <martin.peres@linux.intel.com>
Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>