igt/perf_pmu: Bump batch_duration for legacy sampling inaccuracy
Since the legacy ringbuffer uses a sampling technique, it is limited to
an accuracy based on a 200Hz timer, or 5ms. We assert that measurements
are within 5%, so with a 100ms duration that gives us no room for the
systematic error in our sampling. Bump the duration to 500ms to give us
plenty of safety margin (back of the paper estimate is that we need
300ms to provide the accuracy we request for legacy submission), if it
then fails, it should not be due to the sampling!

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>