igt: Remove gem_ctx_basic
This is just a very plain stress test that doesn't do any verification,
and is entirely duplicated by the other context tests. The test
currently leaks objects from every thread on every pass (which is
triggering an oom on smaller machines), while fixing it would be trivial,
it also is pointless as the test is of little merit.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>