igt/drv_selftests: Disable initialising the display
Many of the selftests try to completely fill global resources; resources
that are presumed available for bringing up the display. Avoid the
contention by simply not bringing up the display!

This does limit the effectiveness of selftesting to GEM for the
time being. To exercise KMS from selftests we would essentially have to
always mock the displays.

References: https://bugs.freedesktop.org/show_bug.cgi?id=103718
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>