tests/gem_exec_schedule: Add reset on failed preemption test.
This patch adds a test where a low priority batch is going to be
declared hung while a preemption is pending. The test wants to verify
that a 'hanging' low priority batch will not disrupt the execution of a high
priority context and that the driver does due diligence in managing a
reset while a preemption is pending.

-v2:
	- Use igt_hang_ctx to hang the engine (Chris)
	- Enable/Process engine reset using IGT libs (Chris)
	- Create new subtest_group for the test (Chris)

Cc: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Michal Winiarski <michal.winiarski@intel.com>
Signed-off-by: Antonio Argenziano <antonio.argenziano@intel.com>
Reviewed-by: Michał Winiarski <michal.winiarski@intel.com>