tests/perf_pmu: Exercise busy stats and lite-restore
While developing a fix for an accounting hole in busy stats we realized
lite-restore is a potential edge case which would be interesting to check
is properly handled.

It is unfortnately quite timing sensitive to hit lite-restore in the
fashion test needs, so downside of this test is that it sufferes from a
high rate of false negatives.

v2:
 * Make the sleep unconditional and use scientific notiation for large
   constants. (Chris Wilson)
 * Use gem_quiscent_gpu instead of gem_sync+usleep to ensure context
   complete was received under execlists. (Chris Wilson)

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>