igt/pm_rps: Increase load for waitboosting
The waitboost subtests encode internal knowledge of the kernel, and
should we wish to change how the kernel functions, we also need to
update the test (or reconfigure it somehow to express desired latencies
for certain workloads). One such change proposed is to remove the
waitboost if the target request is also completed, so update the test to
not wait on the current batch.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>