meson: Name the project intel-gpu-tools
Eventually we're switching the official name to "IGT GPU Tools", but
right now there's still a lot of hardcoding to intel-gpu-tools that is
to be fixed in the near future.

Rename the project in toplevel meson.build to intel-gpu-tools to get
meson to generate tarballs roughly the same as autotools in 'dist'.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>