tests/kms_rotation_crc: Test all pixel formats on all planes.
The test is modified to test all pixel formats on a plane, unless
the pixel format is overridden by the subtest.

Attempting to test all pixel formats requires even more runtime and
the overhead of toggling crc collection to capture a single CRC
becomes significant, so keep the crc collection enabled during the entire test.

This reduces the runtime from ~24s to ~16s per subtest on SKL.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Mika Kahola <mika.kahola@intel.com>
Tested-by: Vidya Srinivas <vidya.srinivas@intel.com>